Split climate_change.rs (807 lines) into subdirectory modules
Break the monolithic climate_change.rs into focused submodules:
- constants.rs (39 lines): CO2 rates, thresholds, tuning parameters
- state.rs (61 lines): ClimateState resource, Default, Saveable impl
- calculations.rs (106 lines): pure helper functions for temperature, scoring
- systems.rs (120 lines): yearly_climate_assessment Bevy system
- tests.rs (440 lines): all unit tests
- mod.rs (58 lines): re-exports all pub items for backward compatibility

Part of #1389
